[SOIN] Extrais les résultats de l’organisation
...dans un nouveau composant
